+++ packages/tswebextension/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
 
+## [Unreleased]
+
+### Added
+- Allowlist wildcard support [#2020](https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2020).
+
+
 ## [0.4.4] - 2023-11-13
 
 ### Changed